+++ static/js/humhub/humhub.ui.additions.js
@@ -253,10 +253,20 @@ humhub.module('ui.additions', function (module, require, $) {
             options = {};
         }
 
-        node = (node instanceof $) ? node[0] : node;
-
+        var $node = $(node);
+        node = $node[0];
         var observer = new MutationObserver(function (mutations) {
-            module.applyTo(node);
+            mutations.forEach(function(mutation) {
+                var $nodes = $(mutation.addedNodes).filter(function () {
+                    return this.nodeType === 1; // filter out text nodes
+                });
+
+                $nodes.each(function() {
+                    var $this = $(this);
+                    console.log('Apply additions to ', this);
+                    module.applyTo($this);
+                })
+            });
         });
 
         observer.observe(node, {childList: true, subtree: true});