+++ protected/humhub/docs/CHANGELOG.md
@@ -8,6 +8,7 @@ HumHub Change Log
 - Enh: Added Registraion::EVENT_AFTER_REGISTRATION UserEvent
 - Enh: Added grunt `migrate-up` and `migrate-create` task
 - Enh: Added profile field type `CheckboxList`
+- Fix: Fixed `ui.addition` `MutationObserver`, only apply additions to inserted nodes.
 
 1.2.0 (April 16, 2017)
 --------------------------------